services/ocr-unified/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
paddleocr==2.7.3
paddlepaddle==2.6.1
python-multipart==0.0.6
pillow==10.2.0
python-dateutil==2.8.2
pydantic==2.5.3
